Change role_id type in user_role_resource table to SMALLINT
The role_id column type has been changed from INTEGER to SMALLINT to optimize storage and better reflect the expected range of role identifiers. This should improve database performance and maintain consistency with the roles table.

+++ week-7/project/schema_ddl/011-user_role_resource.sql
@@ -3,9 +3,9 @@ BEGIN;
 DROP TABLE IF EXISTS user_role_resource CASCADE;
 
 CREATE TABLE IF NOT EXISTS user_role_resource (
-    resource_id INTEGER NOT NULL,
-    user_id     INTEGER NOT NULL,
-    role_id     INTEGER NOT NULL,
+    resource_id INTEGER  NOT NULL,
+    user_id     INTEGER  NOT NULL,
+    role_id     SMALLINT NOT NULL,
     FOREIGN KEY (user_id) REFERENCES users(id)
         ON DELETE CASCADE,
     FOREIGN KEY (role_id) REFERENCES roles(id)
